+++ src/Products/PluggableAuthService/tests/test_MoreCaching.py
@@ -198,7 +198,7 @@ class CachingTests(pastc.PASTestCase):
         self.pas.http_auth.extractCredentials = \
             lambda req: {'login': pastc.user_name,
                          'password': pastc.user_password,
-                         'extra': u'M\344dchen'}
+                         'extra': 'M\344dchen'}
 
         user = self.pas.validate(request)
         self.assertIsNotNone(user)
@@ -225,7 +225,7 @@ class CachingTests(pastc.PASTestCase):
         self.pas.http_auth.extractCredentials = \
             lambda req: {'login': pastc.user_name,
                          'password': pastc.user_password,
-                         'extra': u'M\344dchen'.encode('utf-16')}
+                         'extra': 'M\344dchen'.encode('utf-16')}
 
         user = self.pas.validate(request)
         self.assertIsNotNone(user)